+++ sickrage/core/search.py
@@ -324,20 +324,22 @@ def wantedEpisodes(show, fromDate):
     # check through the list of statuses to see if we want any
     for dbData in sickrage.app.main_db.get_many('tv_episodes', show.indexerid):
         if dbData['season'] > 0 and dbData['airdate'] > fromDate.toordinal():
-            curCompositeStatus = int(dbData["status"] or -1)
-            curStatus, curQuality = Quality.splitCompositeStatus(curCompositeStatus)
-
-            if bestQualities:
-                highestBestQuality = max(allQualities)
-            else:
-                highestBestQuality = 0
+            curStatus, curQuality = Quality.splitCompositeStatus(int(dbData["status"] or -1))
 
             # if we need a better one then say yes
-            if (curStatus in (DOWNLOADED, SNATCHED,
-                              SNATCHED_PROPER) and curQuality < highestBestQuality) or curStatus == WANTED:
-                epObj = show.getEpisode(int(dbData["season"]), int(dbData["episode"]))
-                epObj.wantedQuality = [i for i in allQualities if (i > curQuality and i != Quality.UNKNOWN)]
-                wanted.append(epObj)
+            if curStatus not in (WANTED, DOWNLOADED, SNATCHED, SNATCHED_PROPER):
+                continue
+
+            if curStatus != WANTED:
+                if bestQualities:
+                    if curQuality in bestQualities:
+                        continue
+                elif curQuality in anyQualities:
+                    continue
+
+            epObj = show.getEpisode(int(dbData["season"]), int(dbData["episode"]))
+            epObj.wantedQuality = [i for i in allQualities if (i > curQuality and i != Quality.UNKNOWN)]
+            wanted.append(epObj)
 
     return wanted
 
@@ -362,7 +364,7 @@ def searchForNeededEpisodes():
     return results
 
 
-def searchProviders(show, episodes, manualSearch=False, downCurQuality=False, cacheOnly=False):
+def searchProviders(show, episodes, manualSearch=False, downCurQuality=False, updateCache=True, cacheOnly=False):
     """
     Walk providers for information on shows
 
@@ -383,6 +385,7 @@ def searchProviders(show, episodes, manualSearch=False, downCurQuality=False, ca
     origThreadName = threading.currentThread().getName()
 
     def perform_searches():
+        searchResults = {}
         foundResults = {}
         finalResults = []
 
@@ -418,7 +421,8 @@ def searchProviders(show, episodes, manualSearch=False, downCurQuality=False, ca
                     threading.currentThread().setName(origThreadName + "::[" + providerObj.name + "]")
 
                     # update provider RSS cache
-                    if sickrage.app.config.enable_rss_cache: providerObj.cache.update()
+                    if sickrage.app.config.enable_rss_cache and updateCache:
+                        providerObj.cache.update()
 
                     if len(episodes):
                         if search_mode == 'eponly':
@@ -534,6 +538,7 @@ def searchProviders(show, episodes, manualSearch=False, downCurQuality=False, ca
                         # if not, break it apart and add them as the lowest priority results
                         individualResults = splitNZBResult(bestSeasonResult)
                         for curResult in individualResults:
+                            epNum = -1
                             if len(curResult.episodes) == 1:
                                 epNum = curResult.episodes[0].episode
                             elif len(curResult.episodes) > 1: